+++ src/main/java/roomescape/config/WebMvcConfiguration.java
@@ -4,8 +4,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.preprocessor.CheckAdminRoleInterceptor;
-import roomescape.preprocessor.MemberArgumentResolver;
+import roomescape.filter.CheckAdminRoleInterceptor;
+import roomescape.filter.MemberArgumentResolver;
 
 import java.util.List;
