+++ src/main/java/roomescape/filter/MemberArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.preprocessor;
+package roomescape.filter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
